Refactor the internals of analyze
This removes path generation from string concat, this is instead replaced with os.path.join, which takes care to handle the differences between OS.

Additionally removes os.chdir since this makes it harder to understand what is happening in the script, since the current directory in the loop is important, but difficult to reason.

Additionally splits out the csv, json and markdown file writers into their own files.
(We may want to consider a single interface, but i didn't want to touch that yet).

Lastly avoids multiple reads to the config file by putting it into its own module where the file is cached (via imports).

Tests are still coming, but this is just to highlight the progress so far.